pulling up redundant structs into helpers
