+++ test/parallel_api/numeric/numeric.ops/exclusive_scan_multiplies.pass.cpp
@@ -15,6 +15,7 @@
 
 #include "support/utils.h"
 #include "support/scan_serial_impl.h"
+#include "support/utils_scan.h"
 
 #include <random>
 #include <algorithm>
@@ -22,36 +23,6 @@
 #include <vector>
 
 using namespace TestUtils;
-
-template <typename Type>
-struct test_exclusive_scan_with_binary_op
-{
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size, typename T,
-              typename BinaryOp>
-    std::enable_if_t<!TestUtils::is_reverse_v<Iterator1>>
-    operator()(Policy&& exec, Iterator1 in_first, Iterator1 in_last, Iterator2 out_first, Iterator2 out_last,
-               Iterator3 expected_first, Iterator3 /* expected_last */, Size n, T init, BinaryOp binary_op, T trash)
-    {
-        using namespace std;
-
-        exclusive_scan_serial(in_first, in_last, expected_first, init, binary_op);
-
-        auto orr = exclusive_scan(std::forward<Policy>(exec), in_first, in_last, out_first, init, binary_op);
-
-        EXPECT_TRUE(out_last == orr, "exclusive_scan with binary operator returned wrong iterator");
-        EXPECT_EQ_N(expected_first, out_first, n, "wrong result from exclusive_scan with binary operator");
-        std::fill_n(out_first, n, trash);
-    }
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size, typename T,
-              typename BinaryOp>
-    std::enable_if_t<TestUtils::is_reverse_v<Iterator1>>
-    operator()(Policy&& /* exec */, Iterator1 /* in_first */, Iterator1 /* in_last */, Iterator2 /* out_first */,
-               Iterator2 /* out_last */, Iterator3 /* expected_first */, Iterator3 /* expected_last */, Size /* n */,
-               T /* init */, BinaryOp /* binary_op */, T /* trash */)
-    {
-    }
-};
 
 template <typename T>
 void
